+++ .formatter.exs
@@ -1,6 +1,11 @@
 # Used by "mix format"
 [
-  inputs: ["{mix,.formatter}.exs", "{config,lib,test,federation_compatibility}/**/*.{ex,exs}"],
+  inputs: [
+    "{mix,.formatter}.exs",
+    "{config,lib,test}/**/*.{ex,exs}",
+    "federation_compatibility/{mix,.formatter}.exs",
+    "federation_compatibility/{config,lib,test}/**/*.{ex,exs}"
+  ],
   line_length: 120,
   import_deps: [:absinthe]
 ]