+++ test/support/unmatched.ex
@@ -1,0 +1,7 @@
+defmodule Absinthe.Federation.Fixtures.Unmatched do
+  @type t :: %__MODULE__{
+          id: String.t()
+        }
+
+  defstruct id: ""
+end